Make monitoring-engine real - roadmap step 4
Every router in this engine was previously 100% mock: real, DB-ready
tenant-scoped schema underneath, but every single endpoint returned
hardcoded or fabricated data and never touched the database (confirmed
by reading every router/model in full before starting - not assumed).
This replaces that with real behavior across all 5 routers:

- health.py: real HTTP polling (new app/health_poller.py + a fleet
  service registry in app/service_registry.py, 13 known engines with
  corrected names/ports - the old mock dict had 6 wrong names and was
  missing 3 real engines entirely). A new background asyncio loop
  (app/main.py's lifespan) polls every registered engine's own /health
  on settings.health_check_interval and persists a real HealthCheck row
  per service per pass - this is the fleet's first real self-monitoring,
  the whole reason this step was sequenced before the other 5 mock
  engines. /system and /services now read and aggregate the latest
  persisted checks; /{service_id} does a live on-demand poll. A down or
  unreachable service is now an honest UNHEALTHY record with a real
  error message, never silently reported healthy - verified for real
  against both a real local HTTP server (healthy path) and real DNS
  resolution failures against the fleet's actual container names
  (honest-failure path, since no other engines are running in this dev
  environment).

- alerts.py / incidents.py: real create/acknowledge/resolve/list against
  the alerts/incidents tables, with apply_tenant_context() wiring added
  to tenant_base.py (previously missing here, present in content-engine -
  now consistent). Also fixed Incident.severity from an unvalidated
  free-text String column to the real AlertSeverity enum - the same class
  of bug fixed fleet-wide in autonomy-events last session, found in this
  engine's own schema while reading it for this work. Both routers now
  also surface severity_mapping.py's canonical_severity in their
  responses - that module existed and was tested since last session but
  had no real caller anywhere until now.

- logs.py: real ingestion (previously accepted and discarded every
  POSTed log), real time-windowed listing, real case-insensitive
  substring search against persisted messages.

- performance.py: real recording/listing, and /summary now computes real
  averages from the last hour of data (previously fixed constants like
  avg_cpu_usage: 42.3 regardless of what was actually happening). New
  app/self_metrics.py records this engine's own real CPU/memory/disk via
  psutil on a second background loop - verified against this actual
  machine's real readings, not synthetic. PerformanceMetric.value changed
  Integer -> Float to actually hold real percentages like 62.5, not just
  psutil-shaped in an existing repo, so no migration.

Both background loops are gated by a new ENABLE_BACKGROUND_LOOPS setting
(default true), disabled in tests via conftest.py so the suite doesn't
spend every run polling 13 nonexistent localhost ports.

Every router's test file was rewritten to match - the old tests asserted
on the mock shapes ("alerts.py is 100% mock", etc.), which no longer
describes this engine. New tests exercise real DB writes/reads/filters,
real 404s for unknown IDs, real 422s for invalid enum values, and (for
health) both the real success and real failure polling paths. Full suite
re-run clean: 52/52 passing (up from 38, net-new real coverage, not
padding - previous suite tested mock shapes that no longer exist). App
boots cleanly (`import app.main`) and was smoke-verified standalone
outside pytest too: real poll_all_services() run against all 13 services
(13/13 honest UNHEALTHY, real DNS failures logged), and a real poll
against a live local test server (1/1 HEALTHY, real response captured).

Still needs a real Docker host / the 6200 to confirm the background
loops behave correctly under actual container-to-container networking
(MONITORING_TARGET_HOST override exists for non-Docker bare-metal runs,
unset default relies on Docker Compose container-name DNS resolution,
matching how the other engines already resolve each other) - consistent
with every other roadmap step's standing caveat.

+++ app/models/performance_metric.py
@@ -2,7 +2,7 @@
 Performance metric models
 """
 
-from sqlalchemy import Column, String, Integer, DateTime, ForeignKey, JSON, Enum
+from sqlalchemy import Column, String, Float, DateTime, ForeignKey, JSON, Enum
 from sqlalchemy.dialects.postgresql import UUID
 from datetime import datetime
 import uuid
@@ -32,7 +32,7 @@ class PerformanceMetric(TenantBase, Base):
     metric_name = Column(String(100), nullable=False)
     
     # Value
-    value = Column(Integer, nullable=False)
+    value = Column(Float, nullable=False)
     unit = Column(String(50), nullable=True)
     
     # Context
